+++ deps/v8/src/mips64/regexp-macro-assembler-mips64.h
@@ -90,9 +90,8 @@ class RegExpMacroAssemblerMIPS: public NativeRegExpMacroAssembler {
   // Called from RegExp if the stack-guard is triggered.
   // If the code object is relocated, the return address is fixed before
   // returning.
-  static int CheckStackGuardState(Address* return_address,
-                                  Code* re_code,
-                                  Address re_frame);
+  static int64_t CheckStackGuardState(Address* return_address, Code* re_code,
+                                      Address re_frame);
 
   void print_regexp_frame_constants();
